requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
httpx==0.25.0
openai==1.3.0
pydantic==2.5.0
python-dotenv==1.0.0
